merge dev-android into main

See merge request s08-final/S08P31A102!487

+++ android/hikingdom/app/src/main/java/com/example/hikingdom/data/entities/User.kt
@@ -11,8 +11,8 @@ data class User(
     @SerializedName("nickname") val nickname: String,
     @SerializedName("profileUrl") val profileUrl: String,
     @SerializedName("level") val level: Int,
-    @SerializedName("memberId") val memberId: Int,
-    @SerializedName("clubId") val clubId: Int? = null,
+    @SerializedName("memberId") val memberId: Long,
+    @SerializedName("clubId") val clubId: Long? = null,
 ){
     @PrimaryKey(autoGenerate = true)
     var id: Int = 0
